// src/app/interface/restaurant.d.ts
export interface Restaurant {
  id: string;
  name: string;
  email: string;
  description: string;
  contact: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  open: string;
  close: string;
  image: string;
  type: string;
  rating: number;
  active: boolean;
  tags: string[];
}